feedback 3. GiftRanking.tsx 스타일 파일 분리
