+++ src/App.tsx
@@ -1,9 +1,9 @@
 import PageLayout from '@/components/Layout/PageLayout';
 import NavigationBar from '@/components/NavigationBar';
 import GiftThemeSection from '@/components/GiftTheme';
-import ReceiverSelectBox from './components/RecieverSelectBox';
-import Banner from './components/Banner';
-import GiftRanking from './components/GiftRanking';
+import ReceiverSelectBox from '@/components/RecieverSelectBox';
+import Banner from '@/components/Banner';
+import GiftRanking from '@/components/GiftRanking';
